Fixed correlogram plot

+++ backend/controllers/temporal_controller.py
@@ -233,7 +233,7 @@ class TemporalController:
         granularity: Optional[int] = None,
         diff_order: Optional[int] = None,
         num_lags: Optional[int] = None,
-        alpha: Optional[int] = None,
+        alpha: Optional[float] = None,
         serie: Optional[pd.Series] = None,
     ):
         if serie is None:
@@ -291,7 +291,7 @@ class TemporalController:
         granularity: Optional[int] = None,
         diff_order: Optional[int] = None,
         num_lags: Optional[int] = None,
-        alpha: Optional[int] = None,
+        alpha: Optional[float] = None,
         serie: Optional[pd.Series] = None,
     ):
         if serie is None:
@@ -379,7 +379,7 @@ class TemporalController:
         evolution: Optional[str] = None,
         seasonal: Optional[int] = None,
         num_lags: Optional[int] = None,  # Auto Correlogram and P.A.C plot
-        alpha: Optional[int] = None,
+        alpha: Optional[float] = None,
     ):
         if uf:
             self.df = self.df[self.df["SIGLA_UF"] == uf]
@@ -430,7 +430,7 @@ class TemporalController:
             str
         ] = "additive",  # The type of decomposition ('additive' or 'multiplicative')
         num_lags: Optional[int] = None,  # Auto Correlogram and P.A.C plot
-        alpha: Optional[int] = None,
+        alpha: Optional[float] = None,
     ):
         if uf:
             self.df = self.df[self.df["SIGLA_UF"] == uf]
