+++ backend/routes/temporal_routes.py
@@ -106,7 +106,7 @@ async def correlogram(
     granularity: Optional[int] = None,
     diff_order: Optional[int] = None,
     num_lags: Optional[int] = None,
-    alpha: Optional[int] = None,
+    alpha: Optional[float] = None,
 ):
     try:
         controller = TemporalController()
@@ -135,7 +135,7 @@ async def partial_correlogram(
     granularity: Optional[int] = None,
     diff_order: Optional[int] = None,
     num_lags: Optional[int] = None,
-    alpha: Optional[int] = None,
+    alpha: Optional[float] = None,
 ):
     try:
         controller = TemporalController()
@@ -182,7 +182,7 @@ async def stl_decomposition_data(
     evolution: Optional[str] = None,
     seasonal: Optional[int] = None,
     num_lags: Optional[int] = None,
-    alpha: Optional[int] = None,
+    alpha: Optional[float] = None,
 ):
     try:
         controller = TemporalController()
@@ -204,7 +204,7 @@ async def seasonal_decomposition_data(
     period: Optional[int] = 5,
     model: Optional[str] = "additive",
     num_lags: Optional[int] = None,
-    alpha: Optional[int] = None,
+    alpha: Optional[float] = None,
 ):
     try:
         controller = TemporalController()
